rename some e2e tests
