+++ src/__tests__/browserSuites/push-synchronization-retries.spec.js
@@ -57,7 +57,7 @@ const MILLIS_DESTROY = 700;
  *  0.5 secs: SPLIT_KILL event -> /splitChanges: bad response and network error -> SDK_UPDATE not triggered although cache is updated (killLocally) @TODO review if we should trigger SDK_UPDATE anyway
  *  0.6 secs: MY_SEGMENTS_UPDATE event -> /mySegments/nicolas@split.io: fail twice -> SDK_UPDATE not triggered
  */
-export function testSynchronizationFails(mock, assert) {
+export function testSynchronizationRetries(mock, assert) {
   assert.plan(14);
   mock.reset();
 
@@ -102,14 +102,14 @@ export function testSynchronizationFails(mock, assert) {
 
     setTimeout(() => {
       client.once(client.Event.SDK_UPDATE, () => {
-        assert.fail('SDK_UPDATE event must not be triggered due to fetch fails');
+        assert.fail('SDK_UPDATE event must not be triggered due to fetch failures');
       });
       eventSourceInstance.emitMessage(splitKillMessage);
     }, MILLIS_SPLIT_KILL_EVENT); // send a SPLIT_KILL event with a new changeNumber after 0.5 seconds
 
     setTimeout(() => {
       client.once(client.Event.SDK_UPDATE, () => {
-        assert.fail('SDK_UPDATE event must not be triggered due to fetch fails');
+        assert.fail('SDK_UPDATE event must not be triggered due to fetch failures');
       });
       eventSourceInstance.emitMessage(secondMySegmentsUpdateMessage);
     }, MILLIS_SECOND_MYSEGMENT_UPDATE_EVENT); // send a second MY_SEGMENTS_UPDATE event with a new changeNumber after 0.6 seconds
