+++ src/__tests__/nodeSuites/push-initialization-nopush.spec.js
@@ -9,9 +9,9 @@ import { nearlyEqual } from '../utils';
 import { __setEventSource, __restore } from '../../services/getEventSource/node';
 
 const baseUrls = {
-  sdk: 'https://sdk.push-initialization-fails/api',
-  events: 'https://events.push-initialization-fails/api',
-  auth: 'https://auth.push-initialization-fails/api'
+  sdk: 'https://sdk.push-initialization-nopush/api',
+  events: 'https://events.push-initialization-nopush/api',
+  auth: 'https://auth.push-initialization-nopush/api'
 };
 const config = {
   core: {
