+++ src/__tests__/push/browser.spec.js
@@ -1,9 +1,9 @@
 import tape from 'tape-catch';
 import MockAdapter from 'axios-mock-adapter';
-import { testAuthWithPushDisabled, testAuthWith401, testNoEventSource, testNoBase64Support } from '../browserSuites/push-initialization-fails.spec';
+import { testAuthWithPushDisabled, testAuthWith401, testNoEventSource, testNoBase64Support } from '../browserSuites/push-initialization-nopush.spec';
 import { testAuthRetries, testSSERetries } from '../browserSuites/push-initialization-retries.spec';
 import { testSynchronization } from '../browserSuites/push-synchronization.spec';
-import { testSynchronizationFails } from '../browserSuites/push-synchronization-fails.spec';
+import { testSynchronizationRetries } from '../browserSuites/push-synchronization-retries.spec';
 import { testFallbacking } from '../browserSuites/push-fallbacking.spec';
 import { testRefreshToken } from '../browserSuites/push-refresh-token.spec';
 
@@ -18,12 +18,12 @@ tape('## Browser JS - E2E CI Tests for PUSH ##', function (assert) {
   assert.test('E2E / PUSH initialization: fallback to polling if EventSource is not available', testNoEventSource.bind(null, mock));
   assert.test('E2E / PUSH initialization: fallback to polling if EventSource is not available', testNoBase64Support.bind(null, mock));
 
-  assert.test('E2E / PUSH initialization: auth fails and then success', testAuthRetries.bind(null, mock));
-  assert.test('E2E / PUSH initialization: SSE connection fails and then success', testSSERetries.bind(null, mock));
+  assert.test('E2E / PUSH initialization: auth failures and then success', testAuthRetries.bind(null, mock));
+  assert.test('E2E / PUSH initialization: SSE connection failures and then success', testSSERetries.bind(null, mock));
 
   assert.test('E2E / PUSH synchronization: happy paths', testSynchronization.bind(null, mock));
 
-  assert.test('E2E / PUSH synchronization: fails and corner cases', testSynchronizationFails.bind(null, mock));
+  assert.test('E2E / PUSH synchronization: retries due to failures and corner cases', testSynchronizationRetries.bind(null, mock));
 
   assert.test('E2E / PUSH fallbacking', testFallbacking.bind(null, mock));
