+++ src/__tests__/push/node.spec.js
@@ -1,9 +1,9 @@
 import tape from 'tape-catch';
 import MockAdapter from 'axios-mock-adapter';
-import { testAuthWithPushDisabled, testAuthWith401, testNoEventSource } from '../nodeSuites/push-initialization-fails.spec';
+import { testAuthWithPushDisabled, testAuthWith401, testNoEventSource } from '../nodeSuites/push-initialization-nopush.spec';
 import { testAuthRetries, testSSERetries } from '../nodeSuites/push-initialization-retries.spec';
 import { testSynchronization } from '../nodeSuites/push-synchronization.spec';
-import { testSynchronizationFails } from '../nodeSuites/push-synchronization-fails.spec';
+import { testSynchronizationRetries } from '../nodeSuites/push-synchronization-retries.spec';
 import { testFallbacking } from '../nodeSuites/push-fallbacking.spec';
 import { testRefreshToken } from '../nodeSuites/push-refresh-token.spec';
 
@@ -17,12 +17,12 @@ tape('## Node JS - E2E CI Tests for PUSH ##', async function (assert) {
   assert.test('E2E / PUSH initialization: auth with 401', testAuthWith401.bind(null, mock));
   assert.test('E2E / PUSH initialization: fallback to polling if EventSource is not available', testNoEventSource.bind(null, mock));
 
-  assert.test('E2E / PUSH initialization: auth fails and then success', testAuthRetries.bind(null, mock));
-  assert.test('E2E / PUSH initialization: SSE connection fails and then success', testSSERetries.bind(null, mock));
+  assert.test('E2E / PUSH initialization: auth failures and then success', testAuthRetries.bind(null, mock));
+  assert.test('E2E / PUSH initialization: SSE connection failures and then success', testSSERetries.bind(null, mock));
 
   assert.test('E2E / PUSH synchronization: happy paths', testSynchronization.bind(null, mock));
 
-  assert.test('E2E / PUSH synchronization: fails and corner cases', testSynchronizationFails.bind(null, mock));
+  assert.test('E2E / PUSH synchronization: retries due to failures and corner cases', testSynchronizationRetries.bind(null, mock));
 
   assert.test('E2E / PUSH fallbacking', testFallbacking.bind(null, mock));
